re-organized hooks and versions list components
