+++ src/hooks/dataSources/useVersions.js
@@ -3,6 +3,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import { selectors } from 'reducers';
 import { fetchDataSourceVersions } from 'actions/dataSources';
 
+/** @param {{ name: string }} dataSource */
 export default dataSource => {
   const dispatch = useDispatch();
 
@@ -11,7 +12,7 @@ export default dataSource => {
   );
   useEffect(() => {
     if (!dataSource) return;
-    if (versionsCollection.status === 'IDLE') {
+    if (versionsCollection?.status === 'IDLE') {
       dispatch(fetchDataSourceVersions(dataSource));
     }
   }, [dispatch, versionsCollection, dataSource]);
